Refactor for correspondence with api spec sections
