+++ src/account.js
@@ -1,0 +1,5 @@
+export async function getProfile(dongle_id) {
+  let profile = dongle_id || 'me';
+
+  return request.get(profile + '/');
+}
